backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.0.0
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.41.1
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
gTTS==2.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.3
openai==2.0.0
pillow==10.2.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SpeechRecognition==3.10.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
